\name{penMlogistic-package}
\alias{penMlogistic-package}
\docType{package}
\title{
\packageTitle{penMlogistic}
}
\description{
\packageDescription{Fitting a penalized M-estimator for Logistic Regression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{penMlogistic}
\packageIndices{penMlogistic}
See penMlogistic function
}
\author{
\packageAuthor{Gonzalo Chebi}

Maintainer: \packageMaintainer{Gonzalo Chebi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
